pyproject.toml
[tool.poetry]
name = "tinyrag"
version = "0.0.1"
description = ""
authors = ["Eu-Bin KIM <tlrndk123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
guidance = "^0.0.64"
numpy = "^1.25.1"
rank-bm25 = "^0.2.2"
weaviate-client = "^3.22.1"
unstructured = {extras = ["local-inference"], version = "^0.8.1"}
poppler-utils = "^0.1.0"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
